split remote part handling out of snapp.go into snap_remote.go
